+++ ui/src/components/VNCView/VNCSettingsDialog.tsx
@@ -8,6 +8,7 @@ import ShowDotCursor from './VNCSettingForms/ShowDotCursor'
 import ViewOnly from './VNCSettingForms/ViewOnly'
 import { VNCSettings } from '../../stores/vncSettingsStore'
 import Scaling, { ScalingResize } from './VNCSettingForms/Scaling'
+import { DialogProps } from '@toolpad/core'
 
 
 interface VNCSettingsSaveData extends Omit<VNCSettings, 'showDotCursor' | 'viewOnly' | 'scaling'> {
@@ -17,15 +18,8 @@ interface VNCSettingsSaveData extends Omit<VNCSettings, 'showDotCursor' | 'viewO
   'scaling.resize': ScalingResize
 }
 
-interface VNCSettingsDialog {
-  open: boolean
-  close: ()=>void
-  settingsData: VNCSettings
-  onSettingChange: (settingsData: VNCSettings)=>void
-}
 
-
-export default function VNCSettingsDialog({ open, close, settingsData, onSettingChange }: VNCSettingsDialog) {
+export default function VNCSettingsDialog({ open, onClose, payload }: DialogProps<VNCSettings, null | VNCSettings>) {
   const [reset, setReset] = useState<boolean>(false)
 
   useEffect(() => {
@@ -33,7 +27,7 @@ export default function VNCSettingsDialog({ open, close, settingsData, onSetting
   }, [reset])
 
   function save(data: VNCSettingsSaveData) {
-    onSettingChange({
+    onClose({
       qualityLevel: Number(data.qualityLevel),
       compressionLevel: Number(data.compressionLevel),
       showDotCursor: !!data.showDotCursor,
@@ -43,13 +37,12 @@ export default function VNCSettingsDialog({ open, close, settingsData, onSetting
         resize: data['scaling.resize'],
       },
     })
-    close()
   }
 
   return (
     <Dialog
       open={open}
-      onClose={close}
+      onClose={() => onClose(null)}
       maxWidth="sm"
       fullWidth={true}
       PaperProps={{
@@ -66,7 +59,7 @@ export default function VNCSettingsDialog({ open, close, settingsData, onSetting
       <DialogTitle>VNC Settings</DialogTitle>
       <IconButton
         aria-label="close"
-        onClick={close}
+        onClick={() => onClose(null)}
         sx={(theme) => ({
           position: 'absolute',
           right: 8,
@@ -79,27 +72,27 @@ export default function VNCSettingsDialog({ open, close, settingsData, onSetting
       <DialogContent>
         <Stack spacing={1}>
           <QualityLevel
-            initValue={settingsData.qualityLevel}
+            initValue={payload.qualityLevel}
             reset={reset}
           />
           <Divider />
           <CompressionLevel
-            initValue={settingsData.compressionLevel}
+            initValue={payload.compressionLevel}
             reset={reset}
           />
           <Divider />
           <ShowDotCursor
-            initValue={settingsData.showDotCursor}
+            initValue={payload.showDotCursor}
             reset={reset}
           />
           <Divider />
           <ViewOnly
-            initValue={settingsData.viewOnly}
+            initValue={payload.viewOnly}
             reset={reset}
           />
           <Divider />
           <Scaling
-            initValue={settingsData.scaling}
+            initValue={payload.scaling}
             reset={reset}
           />
           <Divider />