+++ ui/src/components/VNCView/VNCViewSkeleton.tsx
@@ -4,11 +4,7 @@ import BackdropContainer from '../utils/BackdropContainer'
 
 export default function VNCViewSkeleton() {
   return (
-    <BackdropContainer sx={{
-      position: 'absolute',
-      width: '100%',
-      height: '100%',
-    }} open={true}>
+    <BackdropContainer open={true}>
       <Skeleton sx={{
         width: '100%',
         height: '100%',