+++ ui/src/libs/docker/DockerCli.ts
@@ -78,27 +78,32 @@ export default class DockerCli extends DockerCliExec {
     return images.length === 1
   }
 
-  pull(image: string, addStdout: (stdout: string)=>void, onFinish: (exitCode: number)=>void) {
-    this.client.cli.exec('pull', [image], {
-      stream: {
-        onOutput(data) {
-          if (data.stdout) {
-            addStdout(data.stdout)
-          }
+  pull(image: string, addStdout: (stdout: string)=>void) {
+    return new Promise<void>((resolve, reject) => (
+      this.client.cli.exec('pull', [image], {
+        stream: {
+          onOutput(data) {
+            if (data.stdout) {
+              addStdout(data.stdout)
+            }
 
-          if (data.stderr) {
-            throw new Error(data.stderr)
-          }
+            if (data.stderr) {
+              reject(data.stderr)
+            }
+          },
+          onError(error) {
+            reject(error)
+          },
+          onClose(exitCode) {
+            if (exitCode !== 0)
+              return reject(exitCode)
+
+            resolve()
+          },
+          splitOutputLines: true,
         },
-        onError(error) {
-          throw error
-        },
-        onClose(exitCode) {
-          onFinish(exitCode)
-        },
-        splitOutputLines: true,
-      },
-    })
+      })
+    ))
   }
 
   run(image: string, options: CliExecOptions) {