+++ CHANGELOG.md
@@ -5,6 +5,7 @@ History of all released versions and changes
 ## v1.1.0
 - VNC settings for scaling the viewport.
 - Add a way to drag the screen of the vnc viewport.
+- Fix a bug on trying to have 2 connections at the same time.
 
 ## v1.0.2
 - Fix Typos.